tests: add wrappers for wl_test requests
The code in tests is shorter and it's more comfy to
use these wrapperes.

These wrappers are synchronous - they wait until the request
is handled and corresponding event is received. In the tests
the requests were mostly used this way anyway. However, there are
still use cases when we need to use the good old wl_test_... request.

NOTE:
Old code was looking usually like this:

  wl_test_do_something();
  client_roundtrip();

The new code looks like this:

  wl_test_do_something();
  do {
    client_roundtrip()
  } while (got_response);

For weston, it does not change anything. The loop will run only once.
But because this approach is more generic, then this way we can
run the test not only under weston, but even under different
compositors.