tests-helper: distill the functionality from handlers
Split out the real code from event handlers (like pointer_handle_motion,
etc.) and put it into standalone function. We'll use it in follow-up
patch where we will register handlers for toytoolkit clients and we'll
use the new functions so that the code won't be duplicated.

v2. use 0x%x instead of %x

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>